Use QuicTLS 1.1.1s
In light of Willy's comments on
https://github.com/haproxy/haproxy/issues/1903#issuecomment-1302171732,
we should use an OpenSSL base of 1.1.1, to avoid the massive performance
penalty of 3.x (with HAProxy at least).

So undo the QuicTLS 3.0.7 revert (from 1.1.1q) and upgrade to QuicTLS 1.1.1s

We should hear more details after HAProxyConf on the exact metrics, and
in general in the future about the state of LibreSSL/WolfSSL/etc as
alternatives.